requirements.txt
Django
djangorestframework
drf-spectacular
drf-spectacular-sidecar
django-jazzmin
django-extensions
drf-material
djoser
Pillow
language-tool-python
markdown
django-cors-headers